Adds Theming Support to v4
This is a **backwards compatible** additive change that adds support for
theming in evergreen v4.

* adds additional primitives per-component which can be overridden as
part of the theme
* adds checking of the user's theme object before using builtin defaults
via optional?.chaining
* includes lodash.clonedeep for extending the default theme
* adds theme stories that show custom themes, nested themes, and
overriding an accessor function for custom icons
* updates all typescript definitions
* updates evergreen's memoizer to support objects (for theme)

Includes the following fixups
* fixup of any stories throwing warnings/errors

Theming allows an individual to either shalowly or deeply customize the
evergreen theme depending on their needs. By enabling a user to extend()
the default theme to create a copy, they have the ability to replace
colors, type, fonts, etc as needed all while continuing to benefit from
some of Evergreen's stronger design decisions around the neutral
scale, box-shadows, etc. This is a massive commit, and the changes are
broken down as follows:

* Any calls to `theme.get...` pass the `theme` object in as the last
variable. This allows any component using useTheme or withTheme to take
advantage of any primitives assigned as part of the ThemeProvider
* helpers are updated to support the `theme` property, checking for an
appropriate section in the theme before falling back to evergreen
defaults. Since we don't use idx or babel's existential operator yet,
this is solved with the pattern `theme && theme.___ ? theme.___ : ___`
where the blanks refer to the property we need
* There is a new `extend()` method added to the default theme, which
uses lodash.deepclone to create a copy of itself.
* A story was created to demonstrate using a custom theme, with only the
default textoverridden.

Increases bundle size to a 220kb maximum to account for theming support