[Dialog] Add support for custom headers and footers (#746)
The `Dialog` component wraps the `children` with two `Pane`'s -- one that you can supply custom props to, and another that you cannot. Because of that second pane, you can't supply a custom header and footer via the body and still have those header/footer stick to the top/bottom of the Dialog (CMIIW!).

We can't just remove that `Pane`, since it would break the rendering behavior of children in a `Dialog`. As a compromise, this PR introduces a `header` and `footer` component that can be used to override the default header and footer.

| Before | After |
| --- | --- |
| Scrolling doesn't work: ![image](https://user-images.githubusercontent.com/2907397/74979321-43e53d80-53e3-11ea-8672-ae099b8ccf81.png) |  Scrolling works: ![image](https://user-images.githubusercontent.com/2907397/74979328-48115b00-53e3-11ea-961e-97e2eaac59eb.png) |